requirements.txt
# AI-Powered Contract Analysis System - Requirements

# Core Dependencies
langgraph>=0.0.1
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.1

# LLM Providers
google-generativeai>=0.3.0
groq>=0.4.0

# Vector Database (use new official package)
pinecone>=6.0.0

# Embeddings & NLP
sentence-transformers>=2.2.0
torch>=2.0.0

# Document Processing
pypdf>=3.0.0
python-docx>=0.8.11
openpyxl>=3.0.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0
